requirements.txt
click
gitpython